+++ spec/support/matchers/match_to_ip_address.rb
@@ -1,8 +1,7 @@
 # frozen_string_literal: true
 
-# rubocop:disable Style/RescueModifier
 RSpec::Matchers.define(:match_to_ip_address) do
   match do |ip_address|
-    IPAddr.new(ip_address) rescue false
+    IPAddr.new(ip_address) rescue false # rubocop:disable Style/RescueModifier
   end
 end