+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.2] - 2019-11-27
+### Changed
+- `Truemail::Configuration`
+- gem development dependencies
+- linters configs
+
+
 ## [1.4.1] - 2019-11-20
 ### Changed
 - gem development dependencies